Rename 'FrontRow' dirs / modules to 'Freckle'.
