+++ library/Freckle/App/Time.hs
@@ -1,0 +1,10 @@
+-- | Time types and functions for @App@
+module Freckle.App.Time
+  ( Seconds(..)
+  ) where
+
+import Prelude
+
+newtype Seconds = Seconds { unSeconds :: Int }
+  deriving stock (Show, Read)
+  deriving newtype (Eq, Num)
