+++ tests/Main.hs
@@ -5,7 +5,7 @@ where
 
 import Prelude
 
-import FrontRow.App.Test.Hspec.Runner (runParConfig)
+import Freckle.App.Test.Hspec.Runner (runParConfig)
 import qualified Spec
 import Test.Hspec
