+++ weeder.dhall
@@ -1,5 +1,5 @@
 -- This is a library with a weak test suite, so we have to define the the whole
 -- thing as a root until we address that.
-{ roots = [ "^FrontRow\\.App\\..*", "^Main\\.main\$", "^Paths_.*" ]
+{ roots = [ "^Freckle\\.App\\..*", "^Main\\.main\$", "^Paths_.*" ]
 , type-class-roots = True
 }
